status_seq_no as int64 HLC; storage, logstore, and auth cleanup
- status_seq_no widened int32->int64 and assigned as a hybrid logical clock (max(unix_nanos, prev+1)). Self-seeds on a freshly provisioned slave so a rebuilt node resumes above any history the primary retained. BREAKING wire change: deploy the primary before slaves.
- Remove unused user_config_versions table, queries, and generated model.
- Drop environment/name from the secondary deployment_configs schema.
- Slim logstore package down to sub.go (remove base/record/tuple).
- Extract jwt claim helpers into backend/util/jwtu; simplify NewStorageAdapter signature.
- SQLite init/migrations updates.

+++ backend/storage/interface.go
@@ -20,7 +20,7 @@ type SecondaryStore interface {
 	OperatorStore
 	MustWriteDeploymentConfig(ctx context.Context, cfg *apigen.DeploymentConfig)
 	FetchDeploymentStatus(id int32) *apigen.DeploymentStatus
-	FetchDeploymentStatusHistorySince(deploymentID int32, sinceSeqNo int32) []*apigen.DeploymentStatus
+	FetchDeploymentStatusHistorySince(deploymentID int32, sinceSeqNo int64) []*apigen.DeploymentStatus
 	SubscribeDeploymentUpdates(machine string) (chan apigen.DeploymentWithStatus, func())
 }
